refactor(commands): extract governance_status to separate module
Extract governance_status() function from status.py to governance_status.py
to reduce file size from 477 to 368 lines (below 400-line CI threshold).

Changes:
- Create src/vibe3/commands/governance_status.py (124 lines)
- Remove governance_status() from src/vibe3/commands/status.py
- Update cli.py to import from governance_status module
- Update test mocks to patch correct module path
- Fix line length issues in test decorators

Verification:
- All governance_status tests pass (5/5)
- Type checking passes (mypy)
- Lint checks pass (ruff)
- File sizes: status.py=368 lines, governance_status.py=124 lines

Fixes CI LOC ceiling check failure in PR #1695
